+++ lib/boruta/oauth/error.ex
@@ -46,10 +46,6 @@ defmodule Boruta.Oauth.Error do
     %{error | format: nil, redirect_uri: nil}
   end
 
-  def with_format(%Error{} = error, %CodeRequest{redirect_uri: redirect_uri, state: state}) do
-    %{error | format: :query, redirect_uri: redirect_uri, state: state}
-  end
-
   def with_format(%Error{error: :invalid_resource_owner} = error, %HybridRequest{
         redirect_uri: redirect_uri,
         state: state,
@@ -65,8 +61,19 @@ defmodule Boruta.Oauth.Error do
     }
   end
 
-  def with_format(%Error{} = error, %HybridRequest{redirect_uri: redirect_uri, state: state} = request) do
-    %{error | format: response_mode(request), redirect_uri: redirect_uri, state: state}
+  def with_format(%Error{error: :invalid_resource_owner} = error, %CodeRequest{
+        redirect_uri: redirect_uri,
+        state: state,
+        prompt: "none"
+      }) do
+    %{
+      error
+      | error: :login_required,
+        error_description: "User is not logged in.",
+        format: :query,
+        redirect_uri: redirect_uri,
+        state: state
+    }
   end
 
   def with_format(%Error{error: :invalid_resource_owner} = error, %TokenRequest{
@@ -82,6 +89,14 @@ defmodule Boruta.Oauth.Error do
         redirect_uri: redirect_uri,
         state: state
     }
+  end
+
+  def with_format(%Error{} = error, %CodeRequest{redirect_uri: redirect_uri, state: state}) do
+    %{error | format: :query, redirect_uri: redirect_uri, state: state}
+  end
+
+  def with_format(%Error{} = error, %HybridRequest{redirect_uri: redirect_uri, state: state} = request) do
+    %{error | format: response_mode(request), redirect_uri: redirect_uri, state: state}
   end
 
   def with_format(%Error{} = error, %TokenRequest{redirect_uri: redirect_uri, state: state}) do