+++ test/boruta/oauth/integration/authorization_code_grant_test.exs
@@ -162,6 +162,31 @@ defmodule Boruta.OauthTest.AuthorizationCodeGrantTest do
                )
     end
 
+    test "returns an error if user is invalid (prompt=none)", %{client: client} do
+      redirect_uri = List.first(client.redirect_uris)
+
+      assert Oauth.authorize(
+               %Plug.Conn{
+                 query_params: %{
+                   "response_type" => "code",
+                   "client_id" => client.id,
+                   "redirect_uri" => redirect_uri,
+                   "prompt" => "none"
+                 }
+               },
+               %ResourceOwner{sub: nil},
+               ApplicationMock
+             ) ==
+               {:authorize_error,
+                %Error{
+                  format: :query,
+                  error: :login_required,
+                  error_description: "User is not logged in.",
+                  status: :unauthorized,
+                  redirect_uri: redirect_uri
+                }}
+    end
+
     test "returns a code", %{client: client, resource_owner: resource_owner} do
       redirect_uri = List.first(client.redirect_uris)