+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - allow lower case bearer authotization header
+- prompt=none management for authorization code grant requests
 
 ## [2.2.1] - 2022-10-16